+PROFILE

+++ src/components/card.js
@@ -1,5 +1,5 @@
 // @todo: Функция создания карточки
-const createCard = (cardInfo, onDeleteCard, onOpenPreview, onLikeCard) => {
+const createCard = (cardInfo, onDeleteCard, onOpenPreview, onLikeCard, isMine, numberLikes, isLiked) => {
   // @todo: Темплейт карточки
   const cardTemplate = document.querySelector("#card-template").content;
   const cardElement = cardTemplate.querySelector(".card").cloneNode(true);
@@ -8,16 +8,25 @@ const createCard = (cardInfo, onDeleteCard, onOpenPreview, onLikeCard) => {
   const cardTitle = cardElement.querySelector(".card__title");
   const deleteButton = cardElement.querySelector(".card__delete-button");
   const likeButton = cardElement.querySelector(".card__like-button");
+  const likeNumberSpan = cardElement.querySelector(".card__like-number");
 
   cardImage.src = cardInfo.link;
   cardImage.alt = cardInfo.name;
   cardTitle.textContent = cardInfo.name;
-
-  //обработчик на кнопку удаления
-  deleteButton.addEventListener("click", (event) => {
-    onDeleteCard(event.target.closest(".places__item"));
-  });
-  //cardImage.addEventListener("click", (event) => onOpenPreview(event));
+  likeNumberSpan.textContent = numberLikes;
+  
+  if (isMine) {
+    //обработчик на кнопку удаления
+    deleteButton.addEventListener("click", (event) => {
+      onDeleteCard(event.target.closest(".places__item"));
+    })
+  } else {
+    deleteButton.style.display = "none";
+  };
+  
+  if (isLiked) {
+    likeButton.classList.add("card__like-button_is-active");
+  };
   cardImage.addEventListener("click", () => onOpenPreview(cardInfo));
   likeButton.addEventListener("click", (event) => onLikeCard(event));
   return cardElement;
